requirements.txt
python-docx>=0.8.11
pandas>=1.3.0
openpyxl>=3.0.7
Pillow>=9.0.0
numpy>=1.21.0
pytesseract>=0.3.8
python-dateutil>=2.8.2 